Modif utilisation des types standards plus implémentations de bluez envoi paquet

+++ sphero-api/API-src/bluetooth/bluez_adaptor.h
@@ -39,7 +39,7 @@ public:
 	/**
 	 * Permet l'envoi d'un flux de données au destinataire
 	 */
-	virtual int send_data(size_t data_length, uint8_t const * data);
+	virtual ssize_t send_data(size_t data_length, uint8_t const * data);
 
 	/**
 	 * Ferme la connexion
